+++ NEWS.md
@@ -1,6 +1,7 @@
 # aflelo 0.1.1
 
 * Small adjustments to optimisation
+* Add predict_match function
 
 # aflelo 0.1.0